moved exceptions to own package, fixed test folder structure
