+++ src/main/talium/system/templateParser/exeptions/UnexpectedEndOfInputException.java
@@ -1,3 +1,3 @@
-package talium.system.templateParser;
+package talium.system.templateParser.exeptions;
 
 public class UnexpectedEndOfInputException extends IllegalArgumentException {}
